Move `makeHeartbeat` into its own file
